CLI polish
This diff:
1. moves account generation related tests to client_proxy instead of grpc_client, which is only suppose to handle grpc related logic.
2. make ClientProxy::wallet private.
3. SignedTransaction::new_for_test -> craft_signed_transaction_for_client
