+++ client/src/grpc_client.rs
@@ -343,63 +343,3 @@ impl GRPCClient {
             .timeout(std::time::Duration::from_millis(5000))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::client_proxy::{AddressAndIndex, ClientProxy};
-    use config::trusted_peers::TrustedPeersConfigHelpers;
-    use libra_wallet::io_utils;
-    use tempfile::NamedTempFile;
-
-    pub fn generate_accounts_from_wallet(count: usize) -> (ClientProxy, Vec<AddressAndIndex>) {
-        let mut accounts = Vec::new();
-        accounts.reserve(count);
-        let file = NamedTempFile::new().unwrap();
-        let mnemonic_path = file.into_temp_path().to_str().unwrap().to_string();
-        let trust_peer_file = NamedTempFile::new().unwrap();
-        let (_, trust_peer_config) = TrustedPeersConfigHelpers::get_test_config(1, None);
-        let trust_peer_path = trust_peer_file.into_temp_path();
-        trust_peer_config.save_config(&trust_peer_path);
-
-        let val_set_file = trust_peer_path.to_str().unwrap().to_string();
-
-        // We don't need to specify host/port since the client won't be used to connect, only to
-        // generate random accounts
-        let mut client_proxy = ClientProxy::new(
-            "", /* host */
-            "", /* port */
-            &val_set_file,
-            &"",
-            None,
-            Some(mnemonic_path),
-        )
-        .unwrap();
-        for _ in 0..count {
-            accounts.push(client_proxy.create_next_account(&["c"]).unwrap());
-        }
-
-        (client_proxy, accounts)
-    }
-
-    #[test]
-    fn test_generate() {
-        let num = 1;
-        let (_, accounts) = generate_accounts_from_wallet(num);
-        assert_eq!(accounts.len(), num);
-    }
-
-    #[test]
-    fn test_write_recover() {
-        let num = 15;
-        let (client, accounts) = generate_accounts_from_wallet(num);
-        assert_eq!(accounts.len(), num);
-
-        let file = NamedTempFile::new().unwrap();
-        let path = file.into_temp_path();
-        io_utils::write_recovery(&client.wallet, &path).expect("failed to write to file");
-
-        let wallet = io_utils::recover(&path).expect("failed to load from file");
-
-        assert_eq!(client.wallet.mnemonic(), wallet.mnemonic());
-    }
-}
